net: tcp: Close the window when < MSS
When the window gets close to 0, we just close it fully.  This should
prevent SWS, though I might be missing something.  See RFC 813 and 1122.
Its not clear if they meant to slam the window closed or to come up with
some other fake window - probably the latter, but this might suffice.

The driver for this is actually qemu's user networking behavior.  If you
*almost* fill Akaros's receive window (qemu guest), but it is still not
fully closed, qemu will stop sending data.  It seems to happen once the
window is less than the MSS.  Qemu waits 5 seconds, then fills the window
completely.  That sounds like a possibly buggy implementation of the
recommended sender algorithm in RFC 813 (grep 'additional.*enhancement').

Using the MSS as the cutoff for a small packet was recommended by RFC 1122.
Actually, they said min(BUF / 2, MSS), and for practical situations, the
MSS is always less than BUF / 2.

The important thing for us is to treat the window as closed and to update
the window when it opens back up.  Regardless of the < MSS business, if we
happened to have closed the window fully on an ACK, we would never tell the
sender, since we were turning off the acktimer.

Note that the TCP stack doens't get a callback or anything when the window opens
up - if we wanted to, we could set up qio callbacks to update the window
when the user drains the read queue.  That's independent of this change,
but figured I'd mention it.  Also, the whole tcptimer setup is a bit hokey.
The acktimers are put in for 1 'count' (50 / 50, grep MSPTICK), and the
expected value for the first tick is 25ms.  Just from looking at tcpdumps,
I'll see Akaros waiting for about 25-50ms between the window-closing ACK
and the window update.  Presumably we could have opened the window before
that.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>